requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pyperclip>=1.8.2
keyboard>=0.13.5
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
torch>=2.0.0
torchaudio>=2.0.0
